tidy: Clean up integer division warnings.
The clang-tidy "integer division result used by floating point" test
pointed out a number of potential rounding issues.  Fix most of these
by converting to a float/double before the division, wither by a an
explicit cast or by changing the integer divisor to a float divisor.
Fix the scheduler by rewriting the code to be more explicit about
what's happening.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-integer-division.html

+++ mythtv/libs/libmythtv/visualisations/videovisualcircles.cpp
@@ -37,7 +37,8 @@ void VideoVisualCircles::DrawPriv(MythPainter *painter, QPaintDevice* device)
 
 bool VideoVisualCircles::InitialisePriv(void)
 {
-    m_range = (m_area.height() / 2) / (m_scale.range() -10);
+    m_range = (static_cast<double>(m_area.height()) / 2.0)
+        / (m_scale.range() - 10);
     m_scaleFactor = 10.0;
     m_falloff = 1.0;
